찜 추가 및 삭제 DB연동 성공

+++ src/main/java/com/project/mapper/ReviewMapper.java
@@ -14,7 +14,7 @@ import com.project.domain.ReviewDomain;
 public interface ReviewMapper {
 
 	// 리뷰 조회
-	public List<ReviewDomain> getReviewsByStoreId(int storeId);
+	public List<ReviewDomain> getReviewsByStoreId(long storeId);
 
 	// 리뷰 삽입
 	void insertReview(ReviewDomain newReview);
